+++ examples/api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # hyperion-sdk-api-demo
 
+## 0.0.12
+
+### Patch Changes
+
+- fix bignumber exponential
+- Updated dependencies
+  - @hyperionxyz/sdk@0.0.19
+
 ## 0.0.11
 
 ### Patch Changes