+++ packages/bridge/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hyperionxyz/bridge
 
+## 0.0.4
+
+### Patch Changes
+
+- fix bignumber exponential
+
 ## 0.0.3
 
 ### Patch Changes